# docs/swagger.yaml
definitions:
  controllers.UserRequest:
    properties:
      email:
        type: string
      senhaHash:
        type: string
    type: object
  models.User:
    properties:
      ativo:
        type: string
      created_at:
        type: string
      email:
        type: string
      id:
        type: integer
      role:
        type: string
      senha:
        type: string
      updated_at:
        type: string
    type: object
host: localhost:8080
info:
  contact: {}
  description: API documentada com Swagger
  title: API em Go com Gin e Gorm
paths:
  /users:
    get:
      description: Retorna todos os usuários cadastrados
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/models.User'
            type: array
      summary: Lista todos os usuários
      tags:
      - Usuários
    post:
      consumes:
      - application/json
      description: Cria um novo usuário com email e senha hash
      parameters:
      - description: Dados do usuário
        in: body
        name: user
        required: true
        schema:
          $ref: '#/definitions/controllers.UserRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.User'
        "400":
          description: Bad Request
          schema:
            additionalProperties:
              type: string
            type: object
      summary: Cria um novo usuário
      tags:
      - Usuários
  /users/{id}:
    delete:
      description: Remove um usuário do banco de dados
      parameters:
      - description: ID do usuário
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties:
              type: string
            type: object
        "400":
          description: Bad Request
          schema:
            additionalProperties:
              type: string
            type: object
      summary: Deleta um usuário pelo ID
      tags:
      - Usuários
    get:
      description: Retorna um usuário específico com base no seu ID
      parameters:
      - description: ID do usuário
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.User'
        "400":
          description: Bad Request
          schema:
            additionalProperties:
              type: string
            type: object
      summary: Busca um usuário pelo ID
      tags:
      - Usuários
    put:
      consumes:
      - application/json
      description: Atualiza os dados (email e senha) de um usuário específico
      parameters:
      - description: ID do usuário
        in: path
        name: id
        required: true
        type: integer
      - description: Novos dados do usuário
        in: body
        name: user
        required: true
        schema:
          $ref: '#/definitions/controllers.UserRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.User'
        "400":
          description: Bad Request
          schema:
            additionalProperties:
              type: string
            type: object
      summary: Atualiza um usuário pelo ID
      tags:
      - Usuários
swagger: "2.0"